Fixed conflict with GLOBAL variables
While using Node-Roles I was also using the GLOBAL variable app, which Node-Roles accidentally overrides inside it's import method by not defining `var app`. I've fixed it in 2 places for both the variable app and prof, assuming the same would hold true for that variable as well. I have tested this change within my app and the conflict is resolved and the roles plugin still functions.